requirements.txt
torch
torchvision
tqdm
numpy
pillow
calflops
scikit-learn
timm
